requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
firebase-admin==6.4.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
pydantic==2.6.0
python-multipart==0.0.9
requests==2.31.0
google-generativeai
